requirements.txt
appnope==0.1.2
asgiref==3.5.2
asttokens==2.0.5
backcall==0.2.0
certifi==2022.5.18.1
debugpy==1.5.1
decorator==5.1.1
distlib==0.3.4
Django==4.0.4
djangorestframework==3.13.1
entrypoints==0.4
executing==0.8.3
filelock==3.6.0
ipykernel==6.9.1
ipython==8.1.1
jedi==0.18.1
jupyter-client==7.1.2
jupyter-core==4.9.2
matplotlib-inline==0.1.3
nest-asyncio==1.5.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
platformdirs==2.5.1
prompt-toolkit==3.0.28
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.11.2
python-dateutil==2.8.2
pytz==2022.1
pyzmq==22.3.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
tornado==6.1
traitlets==5.1.1
virtualenv==20.13.3
wcwidth==0.2.5

knox~=0.1.14